remove duplicated code for polygon fetchers
