+++ src/services/PolygonIOCryptoPriceService.ts
@@ -5,9 +5,11 @@ import {Logger} from 'winston';
 import PriceAggregator from './PriceAggregator.js';
 import Settings from '../types/Settings.js';
 import TimeService from './TimeService.js';
-import PolygonIOCryptoSnapshotFetcher, {SnapshotResponse} from './fetchers/PolygonIOCryptoSnapshotFetcher.js';
-import PolygonIOSingleCryptoPriceFetcher, {SinglePriceResponse} from './fetchers/PolygonIOSingleCryptoPriceFetcher.js';
+import PolygonIOCryptoSnapshotFetcher from './fetchers/PolygonIOCryptoSnapshotFetcher.js';
+import PolygonIOSingleCryptoPriceFetcher from './fetchers/PolygonIOSingleCryptoPriceFetcher.js';
 import {Pair} from '../types/Feed.js';
+import {SinglePriceResponse} from './fetchers/BasePolygonIOSingleFetcher.js';
+import {SnapshotResponse} from './fetchers/BasePolygonIOSnapshotFetcher.js';
 
 @injectable()
 class PolygonIOCryptoPriceService {
