+++ src/services/fetchers/PolygonIOSingleCryptoPriceFetcher.ts
@@ -1,46 +1,21 @@
-import axios from 'axios';
 import {inject, injectable} from 'inversify';
-import {JSONPath} from 'jsonpath-plus';
 
 import Settings from '../../types/Settings.js';
+import {BasePolygonIOSingleFetcher, SinglePriceResponse} from './BasePolygonIOSingleFetcher.js';
 
 @injectable()
-class PolygonIOSingleCryptoPriceFetcher {
-  private apiKey: string;
-  private timeout: number;
-
+class PolygonIOSingleCryptoPriceFetcher extends BasePolygonIOSingleFetcher {
   constructor(@inject('Settings') settings: Settings) {
+    super();
     this.apiKey = settings.api.polygonIO.apiKey;
     this.timeout = settings.api.polygonIO.timeout;
+    this.valuePath = '$.last.price';
   }
 
   async apply({fsym, tsym}: {fsym: string; tsym: string}, raw = false): Promise<SinglePriceResponse | number> {
     const sourceUrl = `https://api.polygon.io/v1/last/crypto/${fsym}/${tsym}?apiKey=${this.apiKey}`;
-
-    const response = await axios.get(sourceUrl, {
-      timeout: this.timeout,
-      timeoutErrorMessage: `Timeout exceeded: ${sourceUrl}`,
-    });
-
-    if (response.status !== 200) {
-      throw new Error(response.data);
-    }
-
-    return raw ? (response.data as SinglePriceResponse) : this.extractValue(response.data, '$.last.price');
+    return this.fetch(sourceUrl, raw);
   }
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  private extractValue = (data: any, valuePath: string): number => {
-    return JSONPath({json: data, path: valuePath})[0];
-  };
-}
-
-export interface SinglePriceResponse {
-  symbol: string;
-  last: {
-    price: number;
-    timestamp: number;
-  };
 }
 
 export default PolygonIOSingleCryptoPriceFetcher;
